requirements.txt
litellm
python-dotenv
platformdirs